+++ playground/src/pages/multiple.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts"></script>
 <template>
-  <DebugUI />
+  <Suspense>
+    <MultipleCanvas />
+  </Suspense>
 </template>